enable a GDRIVE_FOLDER secret to choose the google drive folder to push (#1)
The GDRIVE_FOLDER should be set to OpenDisclosure for the production
repo and it currently defaults to `netfile_redacted`, which is what it
should be for the development repo. The instructions have been updated
to match the currently expected setup, including replace the
`ChenglimEar` organization with `caciviclab`.

Also, I've added redaction configuration for the additional json files
in netfile_samples.