Use gdrive-datastore (#4)
The pypi package gdrive-datastore is now a central place for the code
that allows storing our data files on Google Drive. In this pull
request, we switch over to a pip install for this code and remove it
from the repository. Additionally, dev container support has been added
to ensure the required development environment is completely specified.